% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flir_matrix.R
\name{get_flir_data_matrix}
\alias{get_flir_data_matrix}
\title{Get FLIR Data Matrix}
\usage{
get_flir_data_matrix(flir_csv_path)
}
\arguments{
\item{flir_csv_path}{The path to the FLIR CSV file containing the temperature data.}
}
\value{
A data matrix containing the temperature measurements from the FLIR image, along with the FLIR image ID.
}
\description{
This function retrieves the data matrix that stores the temperature measurements from a FLIR thermal image generated by a FLIR camera.
}
\keyword{FLIR}
\keyword{data}
\keyword{matrix}
\keyword{temperature}
