% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flir_img_id.R
\name{get_flir_img_id}
\alias{get_flir_img_id}
\title{Get FLIR Image ID}
\usage{
get_flir_img_id(flir_obj)
}
\arguments{
\item{flir_csv_path}{The path to the FLIR CSV file.}
}
\value{
The FLIR image number.
}
\description{
This function retrieves the FLIR image number from a CSV file path.
}
\keyword{internal}
