% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flir_parameters.R
\name{get_flir_parameters}
\alias{get_flir_parameters}
\title{Get FLIR Parameters}
\usage{
get_flir_parameters(flir_csv_path)
}
\arguments{
\item{flir_csv_path_path}{A character vector containing the contents of a FLIR CSV file.}
}
\value{
A data frame containing the extracted metadata and parameters.
}
\description{
This function retrieves metadata and parameters from a FLIR image.
}
\keyword{internal}
