% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flir_summary.R
\name{get_flir_summary}
\alias{get_flir_summary}
\title{Summarise FLIR Data}
\usage{
get_flir_summary(flir_obj)
}
\arguments{
\item{flir_obj}{A FLIR object containing temperature data.}
}
\value{
A summary data frame with the calculated statistics for the FLIR data.
}
\description{
This function takes a FLIR object and summarises the data by calculating various statistics such as mean, variance, median, standard deviation, standard error, and interquartile range (IQR).
}
\keyword{FLIR}
\keyword{data}
\keyword{statistics}
\keyword{summarise}
\keyword{summary}
